# Support for disabling DependantChains. Support for bulk dependency Resolution