--- a/krystal-common/src/main/java/com/flipkart/krystal/data/Inputs.java
+++ b/krystal-common/src/main/java/com/flipkart/krystal/data/Inputs.java
@@ -30,7 +30,7 @@ public record Inputs(ImmutableMap<String, InputValue<Object>> values) {
   }
 
   public <T> T getInputValueOrThrow(String inputName) {
-    return this.<T>getInputValueOpt(inputName).orElseThrow();
+    return this.<T>getInputValue(inputName).getValueOrThrow().orElseThrow();
   }
 
   public <T> T getInputValueOrDefault(String inputName, T defaultValue) {