--- a/krystex/src/main/java/com/flipkart/krystal/krystex/ForkJoinExecutorPool.java
+++ b/krystex/src/main/java/com/flipkart/krystal/krystex/ForkJoinExecutorPool.java
@@ -8,6 +8,7 @@ import com.flipkart.krystal.utils.DistributeLeases;
 import com.flipkart.krystal.utils.MultiLeasePool;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.ForkJoinPool;
+import java.util.concurrent.ForkJoinWorkerThread;
 import java.util.concurrent.TimeUnit;
 
 public final class ForkJoinExecutorPool extends MultiLeasePool<ExecutorService> {
@@ -19,7 +20,11 @@ public final class ForkJoinExecutorPool extends MultiLeasePool<ExecutorService> 
             // Default values picked from Executors.newWorkStealingPool(parallelism)
             new ForkJoinPool(
                 1, // Set to 1 because maximumPoolSize is 1 anyway.
-                defaultForkJoinWorkerThreadFactory, // Same as default
+                pool -> {
+                  ForkJoinWorkerThread thread = defaultForkJoinWorkerThreadFactory.newThread(pool);
+                  thread.setName("Krystal-" + thread.getName());
+                  return thread;
+                }, // Same as default
                 null, // Same as default
                 true, // Same as default - used for async event processing.
                 1, // Default is 0. We set to one because we know we will need one thread.