--- a/krystex/src/main/java/com/flipkart/krystal/krystex/KrystalExecutor.java
+++ b/krystex/src/main/java/com/flipkart/krystal/krystex/KrystalExecutor.java
@@ -1,14 +1,14 @@
 package com.flipkart.krystal.krystex;
 
 import com.flipkart.krystal.data.Inputs;
+import com.flipkart.krystal.krystex.node.NodeExecutionConfig;
 import com.flipkart.krystal.krystex.node.NodeId;
 import java.util.concurrent.CompletableFuture;
 
 public interface KrystalExecutor extends AutoCloseable {
 
-  <T> CompletableFuture<T> executeNode(NodeId nodeId, Inputs inputs);
-
-  <T> CompletableFuture<T> executeNode(NodeId nodeId, Inputs inputs, String requestId);
+  <T> CompletableFuture<T> executeNode(
+      NodeId nodeId, Inputs inputs, NodeExecutionConfig executionConfig);
 
   /** Flushes any pending requests. */
   void flush();