--- a/krystex/src/main/java/com/flipkart/krystal/krystex/KrystalExecutor.java
+++ b/krystex/src/main/java/com/flipkart/krystal/krystex/KrystalExecutor.java
@@ -10,7 +10,11 @@ public interface KrystalExecutor extends AutoCloseable {
   <T> CompletableFuture<T> executeNode(
       NodeId nodeId, Inputs inputs, NodeExecutionConfig executionConfig);
 
-  /** Flushes any pending requests. */
+  /**
+   * Flushes any pending requests and waits for all those requests to finish before returning. This
+   * is done so that any subsequent calls to this executor should happen only after the completion
+   * of the pending futures. If this is not done, it will result in unexpected behaviour.
+   */
   void flush();
 
   @Override // to suppress "throws Exception"