--- a/krystex/src/main/java/com/flipkart/krystal/krystex/Logic.java
+++ b/krystex/src/main/java/com/flipkart/krystal/krystex/Logic.java
@@ -1,3 +1,6 @@
 package com.flipkart.krystal.krystex;
 
-public sealed interface Logic permits MainLogic, ResolverLogic {}
+import com.flipkart.krystal.krystex.resolution.MultiResolver;
+import com.flipkart.krystal.krystex.resolution.ResolverLogic;
+
+public sealed interface Logic permits MainLogic, ResolverLogic, MultiResolver {}