--- a/krystex/src/main/java/com/flipkart/krystal/krystex/RequestContext.java
+++ b/krystex/src/main/java/com/flipkart/krystal/krystex/RequestContext.java
@@ -1,3 +1,0 @@
-package com.flipkart.krystal.krystex;
-
-public record RequestContext(Object content) {}