--- a/krystex/src/main/java/com/flipkart/krystal/krystex/RequestId.java
+++ b/krystex/src/main/java/com/flipkart/krystal/krystex/RequestId.java
@@ -1,8 +1,20 @@
 package com.flipkart.krystal.krystex;
 
-public record RequestId(String asString) {
+public record RequestId(String asString, RequestId originatedFrom) {
 
-  public RequestId append(Object suffix) {
-    return new RequestId("%s:%s".formatted(asString, suffix));
+  public RequestId(String requestId) {
+    this(requestId, null);
+  }
+
+  @Override
+  public RequestId originatedFrom() {
+    if (originatedFrom == null) {
+      return this;
+    }
+    return originatedFrom;
+  }
+
+  public RequestId createNewRequest(Object suffix) {
+    return new RequestId("%s:%s".formatted(asString, suffix), originatedFrom());
   }
 }