--- a/krystex/src/main/java/com/flipkart/krystal/krystex/node/DisabledDependantChainException.java
+++ b/krystex/src/main/java/com/flipkart/krystal/krystex/node/DisabledDependantChainException.java
@@ -1,0 +1,7 @@
+package com.flipkart.krystal.krystex.node;
+
+public class DisabledDependantChainException extends RuntimeException {
+  public DisabledDependantChainException(DependantChain disabledDependantChain) {
+    super("The dependant chain %s has been disabled".formatted(disabledDependantChain));
+  }
+}