**Support for disabling DependantChains**
Without this feature, in the past we had to handle recursive dependencies differently. 
Also every skip dependency command was triggering a lot of chatter all across the call graph. Dsiabling dependant chains which we know are not needed will allow us to avoid propagation of skip command and reduce chatter

**Support for bulk dependency Resolution**
Before this change, every resolver was being invoked independently and the resultant inputs were being passed to the dependency node. This means every resolver leads to its own dependency command. This was causing a lot of chatter across the graph especially when a few vajrams were being executed repeatedly in a single executor. To test this we added a new benchmark in `FormulaTest#vajram_benchmark_2()`.

`FormulaTest#vajram_benchmark()` created 1million executors executing Formula vajram once per executor. This test generally finished in < 20 seconds in a 6 core 

`FormulaTest#vajram_benchmark_2()` creates 1000 executors and executes Formula vajram 1000 times per executor - in effect executing Formula vajram 1million times. Before this change, this test case took 6 minutes to complete. After this change this test case finishes in 30 seconds. The hypothesis is that resolving multiple inputs in bulk reduces the chattiness in the graph and increases the throughput.
Evidence in VajramBenchmark2:

Before change:
CommandsQueuedCount: 1,003,999
CommandQueueBypassedCount: 9,998,637

After change:
CommandsQueuedCount: 1,002,000
CommandQueueBypassedCount: 6,003,000


Full numbers:
/*
       Old code performance:
       Total java method time: 29,883,631
       Total java futures time: 66,435,359
       Loop Count: 1,000,000
       Avg. time to Create Executors:14,522 ns
       Avg. time to Enqueue vajrams:2,904 ns
       Avg. time to execute vajrams:190,280 ns
       Throughput executions/s: 5263
       CommandsQueuedCount: 1,003,999
       CommandQueueBypassedCount: 9,998,637
       Platform overhead over native code: 190,251 ns per request
       Platform overhead over reactive code: 190,214 ns per request
       maxActiveLeasesPerObject: 165, peakAvgActiveLeasesPerObject: 164.66666666666666, maxPoolSize: 12
    */
    /*
      Total java method time: 6,624,273
      Total java futures time: 65,579,807
      Loop Count: 1,000,000
      Avg. time to Create Executors:16,842 ns
      Avg. time to Enqueue vajrams:4,125 ns
      Avg. time to execute vajrams:24,266 ns
      Throughput executions/s: 41666
      CommandsQueuedCount: 1,002,000
      CommandQueueBypassedCount: 6,003,000
      Platform overhead over native code: 24,260 ns per request
      Platform overhead over reactive code: 24,201 ns per request
      maxActiveLeasesPerObject: 72, peakAvgActiveLeasesPerObject: 71.33333333333333, maxPoolSize: 12
    */